Don't compare uint with int
Explicitly cast uint to int to fix a warning.

+++ effects/blur/blurshader.cpp
@@ -396,7 +396,7 @@ void ARBBlurShader::unbind()
 #ifndef KWIN_HAVE_OPENGLES
     int boundObject;
     glGetProgramivARB(GL_FRAGMENT_PROGRAM_ARB, GL_PROGRAM_BINDING_ARB, &boundObject);
-    if (boundObject == program) {
+    if (boundObject == (int)program) {
         glBindProgramARB(GL_FRAGMENT_PROGRAM_ARB, 0);
         glDisable(GL_FRAGMENT_PROGRAM_ARB);
     }
